+++ app/components/wallet/BankAccount.stories.tsx
@@ -1,0 +1,13 @@
+import { storiesOf } from "@storybook/react";
+import * as React from "react";
+import { BankAccount } from "./BankAccount";
+
+storiesOf("BankAccount", module).add("default", () => (
+  <BankAccount
+    bankAccount={{ bankAccountNumberLast4: "1234" }}
+    personalData={{
+      lastName: "Last",
+      firstName: "First",
+    }}
+  />
+));
